# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * CastVoice API specification
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface ErrorResponse {
  error: string;
}

export interface AuthUser {
  id: string;
  username: string;
  /** @nullable */
  firstName?: string | null;
  /** @nullable */
  lastName?: string | null;
  /** @nullable */
  profileImage?: string | null;
}

export interface AuthUserResponse {
  isAuthenticated: boolean;
  user?: AuthUser;
}

export interface UserProfile {
  id: number;
  replitUserId: string;
  displayName: string;
  /** @nullable */
  role?: string | null;
  /** @nullable */
  group?: string | null;
  /** @nullable */
  voiceCloneId?: string | null;
  createdAt: string;
}

export interface UpdateUserProfileBody {
  displayName: string;
  /** @nullable */
  role?: string | null;
  /** @nullable */
  group?: string | null;
}

export interface CloneVoiceBody {
  audioDataUrl: string;
  fileName: string;
  displayName: string;
}

export interface CloneVoiceResponse {
  voiceCloneId: string;
}

export type StoryScriptJson = { [key: string]: unknown };

export interface Character {
  id: string;
  name: string;
  description: string;
}

export interface Story {
  id: number;
  title: string;
  genre: string;
  /** @nullable */
  synopsis?: string | null;
  characters: Character[];
  /** @nullable */
  sceneImageUrl?: string | null;
  scriptJson: StoryScriptJson;
  createdAt: string;
}

export interface FetchStoryUrlBody {
  url: string;
}

export interface ParseStoryTextBody {
  text: string;
}

export interface ExtractedStory {
  title: string;
  characters: Character[];
  rawText: string;
}

export type ProjectCastJson = { [key: string]: unknown };

/**
 * @nullable
 */
export type ProjectSceneImagesJson = { [key: string]: unknown } | null;

export interface Project {
  id: number;
  userId: number;
  storyId: number;
  storyTitle: string;
  storyGenre: string;
  castJson: ProjectCastJson;
  status: string;
  /** @nullable */
  finalAudioUrl?: string | null;
  /** @nullable */
  sceneImagesJson?: ProjectSceneImagesJson;
  /** @nullable */
  generationProgress?: number | null;
  /** @nullable */
  generationStep?: string | null;
  createdAt: string;
}

export type CreateProjectBodyCastJson = { [key: string]: unknown };

export interface CreateProjectBody {
  storyId: number;
  castJson: CreateProjectBodyCastJson;
}

export type UpdateProjectBodyCastJson = { [key: string]: unknown };

export interface UpdateProjectBody {
  castJson?: UpdateProjectBodyCastJson;
  status?: string;
}

/**
 * @nullable
 */
export type GenerationStatusSceneImagesJson = { [key: string]: unknown } | null;

export interface GenerationStatus {
  status: string;
  progress: number;
  /** @nullable */
  currentStep?: string | null;
  /** @nullable */
  finalAudioUrl?: string | null;
  /** @nullable */
  sceneImagesJson?: GenerationStatusSceneImagesJson;
}

export interface DesignVoiceBody {
  description: string;
}

export interface DesignVoiceResponse {
  voiceId: string;
}

export interface CreateInviteLinkBody {
  projectId: number;
  characterId: string;
}

export interface InviteLink {
  id: number;
  projectId: number;
  characterId: string;
  uuid: string;
  /** @nullable */
  filledByUserId?: number | null;
  /** @nullable */
  voiceCloneId?: string | null;
  createdAt: string;
}

export interface InviteLinkDetails {
  invite: InviteLink;
  character: Character;
  storyTitle: string;
  isFilled: boolean;
}

export interface SubmitInviteVoiceBody {
  audioDataUrl: string;
  displayName: string;
}