# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: CastVoice API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /projects:
    get:
      operationId: listProjects
      tags: [projects]
      summary: List user's story projects
      responses:
        "200":
          description: List of projects
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Project"
    post:
      operationId: createProject
      tags: [projects]
      summary: Create a new project
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateProjectBody"
      responses:
        "201":
          description: Created project
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Project"
  /projects/{id}:
    get:
      operationId: getProject
      tags: [projects]
      summary: Get a project by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Project
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Project"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    patch:
      operationId: updateProject
      tags: [projects]
      summary: Update project cast or status
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateProjectBody"
      responses:
        "200":
          description: Updated project
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Project"
    delete:
      operationId: deleteProject
      tags: [projects]
      summary: Delete a project
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
  /projects/{id}/generate:
    post:
      operationId: generateAudioDrama
      tags: [projects]
      summary: Start audio drama generation
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "202":
          description: Generation started
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Project"
        "400":
          description: Bad request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /projects/{id}/status:
    get:
      operationId: getGenerationStatus
      tags: [projects]
      summary: Get generation status and progress
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Generation status
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/GenerationStatus"
...
  /invites:
    post:
      operationId: createInviteLink
      tags: [invites]
      summary: Create an invite link for a character
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateInviteLinkBody"
      responses:
        "201":
          description: Created invite link
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/InviteLink"
  /invites/{uuid}:
    get:
      operationId: getInviteLink
      tags: [invites]
      summary: Get an invite link by UUID
      parameters:
        - name: uuid
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Invite link details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/InviteLinkDetails"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /invites/{uuid}/submit:
    post:
      operationId: submitInviteVoice
      tags: [invites]
      summary: Submit voice recording for an invite
      parameters:
        - name: uuid
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SubmitInviteVoiceBody"
      responses:
        "200":
          description: Voice submitted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/InviteLink"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
components:
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
      required:
        - error
...
    Character:
      type: object
      properties:
        id:
          type: string
        name:
          type: string
        description:
          type: string
      required:
        - id
        - name
        - description
...
    Project:
      type: object
      properties:
        id:
          type: integer
        userId:
          type: integer
        storyId:
          type: integer
        storyTitle:
          type: string
        storyGenre:
          type: string
        castJson:
          type: object
        status:
          type: string
        finalAudioUrl:
          type: ["string", "null"]
        sceneImagesJson:
          type: ["object", "null"]
        generationProgress:
          type: ["integer", "null"]
        generationStep:
          type: ["string", "null"]
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - userId
        - storyId
        - storyTitle
        - storyGenre
        - castJson
        - status
        - createdAt
    CreateProjectBody:
      type: object
      properties:
        storyId:
          type: integer
        castJson:
          type: object
      required:
        - storyId
        - castJson
    UpdateProjectBody:
      type: object
      properties:
        castJson:
          type: object
        status:
          type: string
    GenerationStatus:
      type: object
      properties:
        status:
          type: string
        progress:
          type: integer
        currentStep:
          type: ["string", "null"]
        finalAudioUrl:
          type: ["string", "null"]
        sceneImagesJson:
          type: ["object", "null"]
      required:
        - status
        - progress
...
    CreateInviteLinkBody:
      type: object
      properties:
        projectId:
          type: integer
        characterId:
          type: string
      required:
        - projectId
        - characterId
    InviteLink:
      type: object
      properties:
        id:
          type: integer
        projectId:
          type: integer
        characterId:
          type: string
        uuid:
          type: string
        filledByUserId:
          type: ["integer", "null"]
        voiceCloneId:
          type: ["string", "null"]
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - projectId
        - characterId
        - uuid
        - createdAt
    InviteLinkDetails:
      type: object
      properties:
        invite:
          $ref: "#/components/schemas/InviteLink"
        character:
          $ref: "#/components/schemas/Character"
        storyTitle:
          type: string
        isFilled:
          type: boolean
      required:
        - invite
        - character
        - storyTitle
        - isFilled
    SubmitInviteVoiceBody:
      type: object
      properties:
        audioDataUrl:
          type: string
        displayName:
          type: string
      required:
        - audioDataUrl
        - displayName