# lib/api-zod/src/generated/api.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * CastVoice API specification
 * OpenAPI spec version: 0.1.0
 */
import * as zod from "zod";

/**
 * Returns server health status
 * @summary Health check
 */
export const HealthCheckResponse = zod.object({
  status: zod.string(),
});

/**
 * @summary Get current authenticated user
 */
export const GetCurrentAuthUserResponse = zod.object({
  isAuthenticated: zod.boolean(),
  user: zod
    .object({
      id: zod.string(),
      username: zod.string(),
      firstName: zod.string().nullish(),
      lastName: zod.string().nullish(),
      profileImage: zod.string().nullish(),
    })
    .optional(),
});

/**
 * @summary Get current user profile
 */
export const GetUserProfileResponse = zod.object({
  id: zod.number(),
  replitUserId: zod.string(),
  displayName: zod.string(),
  role: zod.string().nullish(),
  group: zod.string().nullish(),
  voiceCloneId: zod.string().nullish(),
  createdAt: zod.coerce.date(),
});

/**
 * @summary Create or update user profile
 */
export const UpdateUserProfileBody = zod.object({
  displayName: zod.string(),
  role: zod.string().nullish(),
  group: zod.string().nullish(),
});

export const UpdateUserProfileResponse = zod.object({
  id: zod.number(),
  replitUserId: zod.string(),
  displayName: zod.string(),
  role: zod.string().nullish(),
  group: zod.string().nullish(),
  voiceCloneId: zod.string().nullish(),
  createdAt: zod.coerce.date(),
});

/**
 * @summary Clone user voice from audio blob URL
 */
export const CloneUserVoiceBody = zod.object({
  audioDataUrl: zod.string(),
  fileName: zod.string(),
  displayName: zod.string(),
});

export const CloneUserVoiceResponse = zod.object({
  voiceCloneId: zod.string(),
});

/**
 * @summary List all available stories
 */
export const ListStoriesResponseItem = zod.object({
  id: zod.number(),
  title: zod.string(),
  genre: zod.string(),
  synopsis: zod.string().nullish(),
  characters: zod.array(
    zod.object({
      id: zod.string(),
      name: zod.string(),
      description: zod.string(),
    }),
  ),
  sceneImageUrl: zod.string().nullish(),
  scriptJson: zod.object({}).passthrough(),
  createdAt: zod.coerce.date(),
});
export const ListStoriesResponse = zod.array(ListStoriesResponseItem);

/**
 * @summary Get a story by ID
 */
export const GetStoryParams = zod.object({
  id: zod.coerce.number(),
});

export const GetStoryResponse = zod.object({
  id: zod.number(),
  title: zod.string(),
  genre: zod.string(),
  synopsis: zod.string().nullish(),
  characters: zod.array(
    zod.object({
      id: zod.string(),
      name: zod.string(),
      description: zod.string(),
    }),
  ),
  sceneImageUrl: zod.string().nullish(),
  scriptJson: zod.object({}).passthrough(),
  createdAt: zod.coerce.date(),
});

/**
 * @summary Fetch and parse a story from a URL
 */
export const FetchStoryFromUrlBody = zod.object({
  url: zod.string(),
});

export const FetchStoryFromUrlResponse = zod.object({
  title: zod.string(),
  characters: zod.array(
    zod.object({
      id: zod.string(),
      name: zod.string(),
      description: zod.string(),
    }),
  ),
  rawText: zod.string(),
});

/**
 * @summary Parse raw text into a story
 */
export const ParseStoryTextBody = zod.object({
  text: zod.string(),
});

export const ParseStoryTextResponse = zod.object({
  title: zod.string(),
  characters: zod.array(
    zod.object({
      id: zod.string(),
      name: zod.string(),
      description: zod.string(),
    }),
  ),
  rawText: zod.string(),
});

/**
 * @summary List user's story projects
 */
export const ListProjectsResponseItem = zod.object({
  id: zod.number(),
  userId: zod.number(),
  storyId: zod.number(),
  storyTitle: zod.string(),
  storyGenre: zod.string(),
  castJson: zod.object({}).passthrough(),
  status: zod.string(),
  finalAudioUrl: zod.string().nullish(),
  sceneImagesJson: zod.object({}).passthrough().nullish(),
  generationProgress: zod.number().nullish(),
  generationStep: zod.string().nullish(),
  createdAt: zod.coerce.date(),
});
export const ListProjectsResponse = zod.array(ListProjectsResponseItem);

/**
 * @summary Create a new project
 */
export const CreateProjectBody = zod.object({
  storyId: zod.number(),
  castJson: zod.object({}).passthrough(),
});

/**
 * @summary Get a project by ID
 */
export const GetProjectParams = zod.object({
  id: zod.coerce.number(),
});

export const GetProjectResponse = zod.object({
  id: zod.number(),
  userId: zod.number(),
  storyId: zod.number(),
  storyTitle: zod.string(),
  storyGenre: zod.string(),
  castJson: zod.object({}).passthrough(),
  status: zod.string(),
  finalAudioUrl: zod.string().nullish(),
  sceneImagesJson: zod.object({}).passthrough().nullish(),
  generationProgress: zod.number().nullish(),
  generationStep: zod.string().nullish(),
  createdAt: zod.coerce.date(),
});

/**
 * @summary Update project cast or status
 */
export const UpdateProjectParams = zod.object({
  id: zod.coerce.number(),
});

export const UpdateProjectBody = zod.object({
  castJson: zod.object({}).passthrough().optional(),
  status: zod.string().optional(),
});

export const UpdateProjectResponse = zod.object({
  id: zod.number(),
  userId: zod.number(),
  storyId: zod.number(),
  storyTitle: zod.string(),
  storyGenre: zod.string(),
  castJson: zod.object({}).passthrough(),
  status: zod.string(),
  finalAudioUrl: zod.string().nullish(),
  sceneImagesJson: zod.object({}).passthrough().nullish(),
  generationProgress: zod.number().nullish(),
  generationStep: zod.string().nullish(),
  createdAt: zod.coerce.date(),
});

/**
 * @summary Delete a project
 */
export const DeleteProjectParams = zod.object({
  id: zod.coerce.number(),
});

/**
 * @summary Start audio drama generation
 */
export const GenerateAudioDramaParams = zod.object({
  id: zod.coerce.number(),
});

/**
 * @summary Get generation status and progress
 */
export const GetGenerationStatusParams = zod.object({
  id: zod.coerce.number(),
});

export const GetGenerationStatusResponse = zod.object({
  status: zod.string(),
  progress: zod.number(),
  currentStep: zod.string().nullish(),
  finalAudioUrl: zod.string().nullish(),
  sceneImagesJson: zod.object({}).passthrough().nullish(),
});

/**
 * @summary Design a voice using ElevenLabs
 */
export const DesignVoiceBody = zod.object({
  description: zod.string(),
});

export const DesignVoiceResponse = zod.object({
  voiceId: zod.string(),
});

/**
 * @summary Create an invite link for a character
 */
export const CreateInviteLinkBody = zod.object({
  projectId: zod.number(),
  characterId: zod.string(),
});

/**
 * @summary Get an invite link by UUID
 */
export const GetInviteLinkParams = zod.object({
  uuid: zod.coerce.string(),
});

export const GetInviteLinkResponse = zod.object({
  invite: zod.object({
    id: zod.number(),
    projectId: zod.number(),
    characterId: zod.string(),
    uuid: zod.string(),
    filledByUserId: zod.number().nullish(),
    voiceCloneId: zod.string().nullish(),
    createdAt: zod.coerce.date(),
  }),
  character: zod.object({
    id: zod.string(),
    name: zod.string(),
    description: zod.string(),
  }),
  storyTitle: zod.string(),
  isFilled: zod.boolean(),
});

/**
 * @summary Submit voice recording for an invite
 */
export const SubmitInviteVoiceParams = zod.object({
  uuid: zod.coerce.string(),
});

export const SubmitInviteVoiceBody = zod.object({
  audioDataUrl: zod.string(),
  displayName: zod.string(),
});

export const SubmitInviteVoiceResponse = zod.object({
  id: zod.number(),
  projectId: zod.number(),
  characterId: zod.string(),
  uuid: zod.string(),
  filledByUserId: zod.number().nullish(),
  voiceCloneId: zod.string().nullish(),
  createdAt: zod.coerce.date(),
});