# lib/api-zod/src/generated/types/createInviteLinkBody.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * CastVoice API specification
 * OpenAPI spec version: 0.1.0
 */

export interface CreateInviteLinkBody {
  projectId: number;
  characterId: string;
}