# lib/api-zod/src/generated/types/generationStatus.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * CastVoice API specification
 * OpenAPI spec version: 0.1.0
 */
import type { GenerationStatusSceneImagesJson } from "./generationStatusSceneImagesJson";

export interface GenerationStatus {
  status: string;
  progress: number;
  /** @nullable */
  currentStep?: string | null;
  /** @nullable */
  finalAudioUrl?: string | null;
  /** @nullable */
  sceneImagesJson?: GenerationStatusSceneImagesJson;
}