# lib/api-zod/src/generated/types/submitInviteVoiceBody.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * CastVoice API specification
 * OpenAPI spec version: 0.1.0
 */

export interface SubmitInviteVoiceBody {
  audioDataUrl: string;
  displayName: string;
}