# lib/api-zod/src/generated/types/updateProjectBody.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * CastVoice API specification
 * OpenAPI spec version: 0.1.0
 */
import type { UpdateProjectBodyCastJson } from "./updateProjectBodyCastJson";

export interface UpdateProjectBody {
  castJson?: UpdateProjectBodyCastJson;
  status?: string;
}